Features- 1.Successfully created relational database for users data and seller profile, customer profile.           2. Relational database for seller and products
