+++ api/routes/customer.js
@@ -1,8 +1,8 @@
 import express from 'express';
 import jwt from 'jsonwebtoken';
-import getAlluser from '../../controller/fetchAllusers.js';
-import getProfileuserId from '../../controller/getProfileuserId.js';
-import newProfile from '../../controller/customerProfile.js';
+import getAlluser from '../../controller/fetchAllusers';
+import getProfileuserId from '../../controller/getProfileuserId';
+import newProfile from '../../controller/customerProfile';
 
 const router = express.Router();
