+++ controller/fetchAllusers.js
@@ -1,4 +1,4 @@
-import User from '../models/userModel.js';
+import User from '../models/allModels';
 
 const getalluser = async (req, res) => {
   try {
