+++ controller/getProductbycategory.js
@@ -1,8 +1,8 @@
-import Product from '../models/productModel.js';
+import {Products} from '../models/allModels';
 
 const getProductbySeller = async (req, res) => {
   try {
-    const products = await Product.findAll({ where: { productCategory: req.query.category } });
+    const products = await Products.findAll({ where: { productCategory: req.query.category } });
     if (products) {
       return res.status(200).send(products);
     }
