+++ controller/getSingleProduct.js
@@ -1,9 +1,9 @@
-import Product from '../models/productModel.js';
+import {Products} from '../models/allModels';
 
 const getSingleProduct = async (req, res) => {
   try {
     console.log(req.query.id);
-    const products = await Product.findAll({ where: { id: req.query.id } });
+    const products = await Products.findAll({ where: { id: req.query.id } });
     if (products) {
       return res.status(200).send(products);
     }
