+++ controller/updateProduct.js
@@ -1,8 +1,8 @@
-import Product from '../models/productModel.js';
+import {Products} from '../models/allModels';
 
 const updateProduct = async (req, res) => {
   try {
-    const product = await Product.update({
+    const product = await Products.update({
       sellerId: req.params.sellerId,
       imgpath: req.body.imgpath,
       productName: req.body.productName,
