+++ controller/userLogin.js
@@ -1,6 +1,6 @@
 import jwt from 'jsonwebtoken';
 import bcrypt from 'bcrypt';
-import User from '../models/userModel.js';
+import User from '../models/allModels';
 
 const loginAuth = async (req, res) => {
   try {
