+++ controller/viewOrderbyCustomer.js
@@ -1,4 +1,4 @@
-import Order from '../models/orderModel.js';
+import {Order} from '../models/allModels';
 
 const viewOrder = async (req, res) => {
   try {
