+++ test/test.js
@@ -1,6 +1,6 @@
 import chai from 'chai';
 import chaiHttp from 'chai-http';
-import app from '../app.js';
+import app from '../app';
 
 import  jwt from 'jsonwebtoken';
